chore(DTFS2-7431): clean bss ewcs deal function
